Moved requests to services into a separate file to facilitate reuse. Allowed interactive batch, or list and force modes
